+++ src/device/tsp.script/Script/SessionBaseExtensions/FirmwareManager-load.cs
@@ -259,7 +259,7 @@ public static partial class FirmwareManager
                 {
                     if ( lineType == LuaChunkLineContentType.SyntaxStartCommentBlock )
                     {
-                        chunkLine = chunkLine[..chunkLine.IndexOf( Syntax.Tsp.Lua.StartCommentChunk, StringComparison.OrdinalIgnoreCase )];
+                        chunkLine = chunkLine[..chunkLine.IndexOf( Syntax.Tsp.Lua.StartCommentBlockChunk, StringComparison.OrdinalIgnoreCase )];
                     }
 
                     // end each line with a space
@@ -394,7 +394,7 @@ public static partial class FirmwareManager
                 {
                     if ( lineType == LuaChunkLineContentType.SyntaxStartCommentBlock )
                     {
-                        chunkLine = chunkLine[..chunkLine.IndexOf( Syntax.Tsp.Lua.StartCommentChunk, StringComparison.OrdinalIgnoreCase )];
+                        chunkLine = chunkLine[..chunkLine.IndexOf( Syntax.Tsp.Lua.StartCommentBlockChunk, StringComparison.OrdinalIgnoreCase )];
                     }
 
                     // end each line with a space
@@ -551,7 +551,7 @@ public static partial class FirmwareManager
             {
                 if ( lineType == LuaChunkLineContentType.SyntaxStartCommentBlock )
                 {
-                    chunkLine = chunkLine[..chunkLine.IndexOf( Syntax.Tsp.Lua.StartCommentChunk, StringComparison.OrdinalIgnoreCase )];
+                    chunkLine = chunkLine[..chunkLine.IndexOf( Syntax.Tsp.Lua.StartCommentBlockChunk, StringComparison.OrdinalIgnoreCase )];
                 }
 
                 // end each line with a space